Add studymeta strategy as a factory
